# docs/api-spec/loyalty-rewards-service-openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: loyalty@probestack.io
    name: Growth Engineering Team
  description: |
    Manages loyalty point balances, tier progression, and reward redemptions.
    This API is used as the source specification to automatically generate
    an API Gateway proxy (Kong or Apigee) with routing, authentication,
    rate limiting, and caching policies.
  title: Loyalty Rewards Service
  version: 1.0.0
  x-proxy-config:
    upstream-url: https://loyalty-service.probestack.io
    service-name: loyalty-rewards-gateway
    rate-limit: 250/minute
    cors-enabled: true
    retries: 2
servers:
- description: Production Gateway
  url: https://gateway.probestack.io/loyalty/v1
- description: Local Gateway
  url: http://localhost:8000/loyalty/v1
security:
- ApiKeyAuth: []
...
paths:
...
  /rewards:
    get:
      description: "Returns the reward catalog, optionally filtered to rewards the\
        \ member can currently afford."
      operationId: listRewards
      parameters:
      - description: Return only rewards this member has enough points for
        explode: true
        in: query
        name: affordableForMemberId
        required: false
        schema:
          format: uuid
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Reward"
                type: array
          description: Rewards fetched successfully
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid API Key.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: List available rewards
      tags:
      - Rewards
      x-proxy-route:
        methods:
        - GET
        plugins:
        - name: cors
        - name: proxy-cache
          config:
            cache_ttl: 180
      x-accepts: application/json
      x-tags:
      - tag: Rewards
  /rewards/{rewardId}/redeem:
    post:
      description: Burns the required points and issues a redemption code. Redemption
        is atomic.
      operationId: redeemReward
      parameters:
      - explode: false
        in: path
        name: rewardId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemRewardRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
          description: Reward redeemed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request payload or parameters.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid API Key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Redeem a reward
      tags:
      - Rewards
      x-proxy-route:
        methods:
        - POST
        plugins:
        - name: cors
        - name: rate-limiting
          config:
            minute: 40
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Rewards
components:
...
  schemas:
...
    Reward:
      example:
        pointCost: 1000
        name: 10 USD store credit
        available: true
        description: Applies as a discount on your next order
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        category: CREDIT
      properties:
        id:
          format: uuid
          type: string
        name:
          example: 10 USD store credit
          type: string
        description:
          example: Applies as a discount on your next order
          type: string
        pointCost:
          example: 1000
          type: integer
        category:
          enum:
          - CREDIT
          - MERCHANDISE
          - EXPERIENCE
          - DONATION
          example: CREDIT
          type: string
        available:
          example: true
          type: boolean
      type: object
    RedeemRewardRequest:
      example:
        deliveryEmail: john.doe@example.com
        memberId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        memberId:
          format: uuid
          type: string
        deliveryEmail:
          example: john.doe@example.com
          format: email
          type: string
      required:
      - memberId
      type: object
    Redemption:
      example:
        redemptionCode: RWD-8H3K-92QP
        rewardId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        pointsBurned: 1000
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        expiresAt: 2000-01-23T04:56:07.000+00:00
        memberId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: ISSUED
      properties:
        id:
          format: uuid
          type: string
        rewardId:
          format: uuid
          type: string
        memberId:
          format: uuid
          type: string
        pointsBurned:
          example: 1000
          type: integer
        redemptionCode:
          example: RWD-8H3K-92QP
          type: string
        status:
          enum:
          - ISSUED
          - USED
          - EXPIRED
          example: ISSUED
          type: string
        expiresAt:
          format: date-time
          type: string
      type: object
    ErrorResponse:
      properties:
        status:
          example: 400
          type: integer
        errorCode:
          example: INSUFFICIENT_POINTS
          type: string
        message:
          example: Member does not have enough points for this reward
          type: string
        timestamp:
          format: date-time
          type: string
        path:
          example: /loyalty/v1/rewards
          type: string
      required:
      - message
      - status
      type: object
  securitySchemes:
    ApiKeyAuth:
      description: API Key required for all gateway requests.
      in: header
      name: X-API-Key
      type: apiKey