# src/main/java/com/probestack/forgestudio/design/api/RewardsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.ErrorResponse;
import com.probestack.forgestudio.design.model.RedeemRewardRequest;
import com.probestack.forgestudio.design.model.Redemption;
import com.probestack.forgestudio.design.model.Reward;
import java.util.UUID;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-13T10:36:55.986768776Z[GMT]")@Validated
@Tag(name = "Rewards", description = "Reward catalog and redemption")
public interface RewardsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /rewards : List available rewards
     * Returns the reward catalog, optionally filtered to rewards the member can currently afford.
     *
     * @param affordableForMemberId Return only rewards this member has enough points for (optional)
     * @return Rewards fetched successfully (status code 200)
     *         or Missing or invalid API Key. (status code 401)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "listRewards",
        summary = "List available rewards",
        description = "Returns the reward catalog, optionally filtered to rewards the member can currently afford.",
        tags = { "Rewards" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Rewards fetched successfully", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Reward.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Missing or invalid API Key.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/rewards",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Reward>> listRewards(
        @Parameter(name = "affordableForMemberId", description = "Return only rewards this member has enough points for", in = ParameterIn.QUERY) @Valid @RequestParam(value = "affordableForMemberId", required = false) UUID affordableForMemberId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "[ { \"pointCost\" : 1000, \"name\" : \"10 USD store credit\", \"available\" : true, \"description\" : \"Applies as a discount on your next order\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"category\" : \"CREDIT\" }, { \"pointCost\" : 1000, \"name\" : \"10 USD store credit\", \"available\" : true, \"description\" : \"Applies as a discount on your next order\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"category\" : \"CREDIT\" } ]";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * POST /rewards/{rewardId}/redeem : Redeem a reward
     * Burns the required points and issues a redemption code. Redemption is atomic.
     *
     * @param rewardId  (required)
     * @param redeemRewardRequest  (required)
     * @return Reward redeemed successfully (status code 201)
     *         or Invalid request payload or parameters. (status code 400)
     *         or Missing or invalid API Key. (status code 401)
     *         or Resource not found. (status code 404)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "redeemReward",
        summary = "Redeem a reward",
        description = "Burns the required points and issues a redemption code. Redemption is atomic.",
        tags = { "Rewards" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Reward redeemed successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Redemption.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing or invalid API Key.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/rewards/{rewardId}/redeem",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Redemption> redeemReward(
        @Parameter(name = "rewardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("rewardId") UUID rewardId,
        @Parameter(name = "RedeemRewardRequest", description = "", required = true) @Valid @RequestBody RedeemRewardRequest redeemRewardRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"redemptionCode\" : \"RWD-8H3K-92QP\", \"rewardId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"pointsBurned\" : 1000, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"expiresAt\" : \"2000-01-23T04:56:07.000+00:00\", \"memberId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"status\" : \"ISSUED\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}